+++ packages/riverpod/lib/src/stream_provider/base.dart
@@ -16,9 +16,8 @@ abstract class StreamProviderRef<State> implements Ref {
 
 /// {@macro riverpod.streamprovider}
 @sealed
-class StreamProvider<State> extends AsyncProvider<State>
+class StreamProvider<State> extends AlwaysAliveProviderBase<AsyncValue<State>>
     with
-        _StreamProviderMixin<State>,
         OverrideWithValueMixin<AsyncValue<State>>,
         OverrideWithProviderMixin<AsyncValue<State>,
             AlwaysAliveProviderBase<AsyncValue<State>>> {
@@ -43,13 +42,127 @@ class StreamProvider<State> extends AsyncProvider<State>
   @override
   final List<ProviderOrFamily>? dependencies;
 
-  @override
+  /// {@template riverpod.streamprovider.stream}
+  /// Exposes the [Stream] created by a [StreamProvider].
+  ///
+  /// The stream obtained is not strictly identical to the stream created.
+  /// Instead, this stream is always a broadcast stream.
+  ///
+  /// The stream obtained may change over time, if the [StreamProvider] is
+  /// re-evaluated, such as when it is using [Ref.watch] and the
+  /// provider listened changes, or on [ProviderContainer.refresh].
+  ///
+  /// If the [StreamProvider] was overridden using `overrideWithValue`,
+  /// a stream will be generated and manipulated based on the [AsyncValue] used.
+  /// {@endtemplate}
   late final AlwaysAliveProviderBase<Stream<State>> stream =
-      AsyncValueAsStreamProvider(this, name: modifierName(name, 'stream'));
+      AsyncValueAsStreamProvider(this);
 
-  @override
-  late final AlwaysAliveProviderBase<Future<State>> last =
-      AsyncValueAsFutureProvider(this, name: modifierName(name, 'last'));
+  /// {@template riverpod.streamprovider.future}
+  /// Exposes a [Future] which resolves with the last value or error emitted.
+  ///
+  /// This can be useful for scenarios where we want to read the current value
+  /// exposed by a [StreamProvider], but also handle the scenario were no
+  /// value were emitted yet:
+  ///
+  /// ```dart
+  /// final configsProvider = StreamProvider<Configuration>((ref) async* {
+  ///   // somehow emit a Configuration instance
+  /// });
+  ///
+  /// final productsProvider = FutureProvider<Products>((ref) async {
+  ///   // If a "Configuration" was emitted, retrieve it.
+  ///   // Otherwise, wait for a Configuration to be emitted.
+  ///   final configs = await ref.watch(configsProvider.last);
+  ///
+  ///   final response = await httpClient.get('${configs.host}/products');
+  ///   return Products.fromJson(response.data);
+  /// });
+  /// ```
+  ///
+  /// ## Why not use [StreamProvider.stream.first] instead?
+  ///
+  /// If you are familiar with streams, you may wonder why not use [Stream.first]
+  /// instead:
+  ///
+  /// ```dart
+  /// final configsProvider = StreamProvider<Configuration>((ref) {...});
+  ///
+  /// final productsProvider = FutureProvider<Products>((ref) async {
+  ///   final configs = await ref.watch(configsProvider.stream).first;
+  ///   ...
+  /// }
+  /// ```
+  ///
+  /// The problem with this code is, unless your [StreamProvider] is creating
+  /// a `BehaviorSubject` from `package:rxdart`, you have a bug.
+  ///
+  /// By default, if we call [Stream.first] **after** the first value was emitted,
+  /// then the [Future] created will not obtain that first value but instead
+  /// wait for a second one – which may never come.
+  ///
+  /// The following code demonstrate this problem:
+  ///
+  /// ```dart
+  /// final exampleProvider = StreamProvider<int>((ref) async* {
+  ///   yield 42;
+  /// });
+  ///
+  /// final anotherProvider = FutureProvider<void>((ref) async {
+  ///   print(await ref.watch(exampleProvider.stream).first);
+  ///   // The code will block here and wait forever
+  ///   print(await ref.watch(exampleProvider.stream).first);
+  ///   print('this code is never reached');
+  /// });
+  ///
+  /// void main() async {
+  ///   final container = ProviderContainer();
+  ///   await container.read(anotherProvider.future);
+  ///   // never reached
+  ///   print('done');
+  /// }
+  /// ```
+  ///
+  /// This snippet will print `42` once, then wait forever.
+  ///
+  /// On the other hand, if we used [StreamProvider.future], our code would
+  /// correctly execute:
+  ///
+  /// ```dart
+  /// final exampleProvider = StreamProvider<int>((ref) async* {
+  ///   yield 42;
+  /// });
+  ///
+  /// final anotherProvider = FutureProvider<void>((ref) async {
+  ///   print(await ref.watch(exampleProvider.future));
+  ///   print(await ref.watch(exampleProvider.future));
+  ///   print('completed');
+  /// });
+  ///
+  /// void main() async {
+  ///   final container = ProviderContainer();
+  ///   await container.read(anotherProvider.future);
+  ///   print('done');
+  /// }
+  /// ```
+  ///
+  /// with this modification, our code will now print:
+  ///
+  /// ```
+  /// 42
+  /// 42
+  /// completed
+  /// done
+  /// ```
+  ///
+  /// which is the expected behavior.
+  /// {@endtemplate}
+  late final AlwaysAliveProviderBase<Future<State>> future =
+      AsyncValueAsFutureProvider(this);
+
+  /// {@template riverpod.streamprovider.future}
+  @Deprecated('use `future` instead')
+  AlwaysAliveProviderBase<Future<State>> get last => future;
 
   @override
   AsyncValue<State> create(covariant StreamProviderElement<State> ref) {
@@ -74,7 +187,8 @@ class StreamProvider<State> extends AsyncProvider<State>
 }
 
 /// The Element of a [StreamProvider]
-class StreamProviderElement<State> extends AsyncProviderElement<State>
+class StreamProviderElement<State>
+    extends ProviderElementBase<AsyncValue<State>>
     with _StreamProviderElementMixin<State>
     implements StreamProviderRef<State> {
   /// The Element of a [StreamProvider]
@@ -84,16 +198,7 @@ class StreamProviderElement<State> extends AsyncProviderElement<State>
   AsyncValue<State> get state => requireState;
 
   @override
-  set state(AsyncValue<State> newState) {
-    assert(
-      newState is AsyncData ||
-          (newState is AsyncLoading &&
-              (newState as AsyncLoading).previous == null) ||
-          (newState is AsyncError && (newState as AsyncError).previous == null),
-      'Cannot specify "previous" for AsyncValue but got $newState',
-    );
-    setState(newState);
-  }
+  set state(AsyncValue<State> newState) => setState(newState);
 }
 
 /// {@template riverpod.streamprovider.family}
@@ -118,9 +223,6 @@ class StreamProviderFamily<State, Arg>
       name: name,
     );
 
-    registerProvider(provider.stream, argument);
-    registerProvider(provider.last, argument);
-
     return provider;
   }
 
@@ -128,8 +230,6 @@ class StreamProviderFamily<State, Arg>
   void setupOverride(Arg argument, SetupOverride setup) {
     final provider = call(argument);
     setup(origin: provider, override: provider);
-    setup(origin: provider.stream, override: provider.stream);
-    setup(origin: provider.last, override: provider.last);
   }
 
   /// {@macro riverpod.overridewithprovider}
@@ -139,8 +239,6 @@ class StreamProviderFamily<State, Arg>
     return FamilyOverride<Arg>(this, (arg, setup) {
       final provider = call(arg);
       setup(origin: provider, override: override(arg));
-      setup(origin: provider.stream, override: provider.stream);
-      setup(origin: provider.last, override: provider.last);
     });
   }
 }