+++ packages/riverpod/test/providers/stream_provider/stream_provider_test.dart
@@ -31,7 +31,7 @@ void main() {
 
     container.listen(provider, listener);
 
-    await container.read(provider.last);
+    await container.read(provider.future);
     expect(ref.state, const AsyncData<int>(0));
     verifyOnly(
       listener,
@@ -45,39 +45,15 @@ void main() {
 
     expect(
       ref.state,
-      const AsyncLoading<int>(previous: AsyncData(0)),
+      const AsyncLoading<int>(),
     );
 
     verifyOnly(
       listener,
       listener(
         const AsyncData(0),
-        const AsyncLoading<int>(previous: AsyncData(0)),
+        const AsyncLoading<int>(),
       ),
-    );
-  });
-
-  test('throws if trying to set AsyncValue with a previous value', () {
-    final container = createContainer();
-    final listener = Listener<AsyncValue<int>>();
-    late StreamProviderRef<int> ref;
-    final provider = StreamProvider<int>((r) {
-      ref = r;
-      return Stream.value(0);
-    });
-
-    container.listen(provider, listener);
-
-    expect(ref.state, const AsyncLoading<int>());
-    verifyZeroInteractions(listener);
-
-    expect(
-      () => ref.state = const AsyncLoading<int>(previous: AsyncData(42)),
-      throwsA(isA<AssertionError>()),
-    );
-    expect(
-      () => ref.state = const AsyncError<int>(42, previous: AsyncData(42)),
-      throwsA(isA<AssertionError>()),
     );
   });
 
@@ -94,7 +70,7 @@ void main() {
     );
 
     await expectLater(container.read(provider.stream), emits(42));
-    await expectLater(container.read(provider.last), completion(42));
+    await expectLater(container.read(provider.future), completion(42));
     expect(container.read(provider), const AsyncData(42));
 
     expect(root.getAllProviderElements(), isEmpty);
@@ -125,7 +101,7 @@ void main() {
 
     verifyOnly(
       listener,
-      listener(null, const AsyncLoading<int>(previous: AsyncData(42))),
+      listener(null, const AsyncLoading<int>()),
     );
 
     container.read(dep).state = Stream.value(21);
@@ -142,83 +118,23 @@ void main() {
     );
   });
 
-  test(
-      'when recreating the future, AsyncLoading contains the previous data if any',
-      () async {
-    final dep = StateProvider((ref) => Stream.value(42));
-    final provider = StreamProvider((ref) => ref.watch(dep).state);
-    final container = createContainer();
-
-    expect(
-      container.read(provider),
-      const AsyncLoading<int>(),
-    );
-
-    await expectLater(
-      container.read(provider.stream),
-      emits(42),
-    );
-    expect(
-      container.read(provider),
-      const AsyncData<int>(42),
-    );
-
-    container.read(dep).state = Stream.value(21);
-
-    expect(
-      container.read(provider),
-      const AsyncLoading<int>(previous: AsyncData(42)),
-    );
-  });
-
-  test(
-      'when recreating the future, AsyncLoading contains the previous error if any',
-      () async {
-    final dep = StateProvider((ref) => Stream<int>.error(42, StackTrace.empty));
-    final provider = StreamProvider((ref) => ref.watch(dep).state);
-    final container = createContainer();
-
-    expect(
-      container.read(provider),
-      const AsyncLoading<int>(),
-    );
-
-    await expectLater(
-      container.read(provider.stream),
-      emitsError(42),
-    );
-    expect(
-      container.read(provider),
-      const AsyncError<int>(42, stackTrace: StackTrace.empty),
-    );
-
-    container.read(dep).state = Stream.value(21);
-
-    expect(
-      container.read(provider),
-      const AsyncLoading<int>(
-        previous: AsyncError(42, stackTrace: StackTrace.empty),
-      ),
-    );
-  });
-
   test('can be refreshed', () async {
     var result = 0;
     final container = createContainer();
     final provider = StreamProvider((ref) => Stream.value(result));
 
     expect(container.read(provider.stream), emits(0));
-    expect(await container.read(provider.last), 0);
+    expect(await container.read(provider.future), 0);
     expect(container.read(provider), const AsyncValue.data(0));
 
     result = 1;
     expect(
       container.refresh(provider),
-      const AsyncValue<int>.loading(previous: AsyncData(0)),
+      const AsyncValue<int>.loading(),
     );
 
     expect(container.read(provider.stream), emits(1));
-    expect(await container.read(provider.last), 1);
+    expect(await container.read(provider.future), 1);
     expect(container.read(provider), const AsyncValue.data(1));
   });
 
@@ -229,7 +145,7 @@ void main() {
       final container = createContainer(parent: root, overrides: [provider]);
 
       expect(await container.read(provider.stream).first, 0);
-      expect(await container.read(provider.last), 0);
+      expect(await container.read(provider.future), 0);
       expect(container.read(provider), const AsyncValue.data(0));
       expect(root.getAllProviderElements(), isEmpty);
       expect(
@@ -238,7 +154,7 @@ void main() {
           isA<ProviderElementBase>()
               .having((e) => e.origin, 'origin', provider),
           isA<ProviderElementBase>()
-              .having((e) => e.origin, 'origin', provider.last),
+              .having((e) => e.origin, 'origin', provider.future),
           isA<ProviderElementBase>()
               .having((e) => e.origin, 'origin', provider.stream),
         ]),
@@ -253,7 +169,7 @@ void main() {
       ]);
 
       expect(await container.read(provider.stream).first, 42);
-      expect(await container.read(provider.last), 42);
+      expect(await container.read(provider.future), 42);
       expect(container.read(provider), const AsyncValue.data(42));
       expect(root.getAllProviderElements(), isEmpty);
       expect(
@@ -262,7 +178,7 @@ void main() {
           isA<ProviderElementBase>()
               .having((e) => e.origin, 'origin', provider),
           isA<ProviderElementBase>()
-              .having((e) => e.origin, 'origin', provider.last),
+              .having((e) => e.origin, 'origin', provider.future),
           isA<ProviderElementBase>()
               .having((e) => e.origin, 'origin', provider.stream),
         ]),
@@ -277,7 +193,7 @@ void main() {
       ]);
 
       expect(await container.read(provider.stream).first, 42);
-      expect(await container.read(provider.last), 42);
+      expect(await container.read(provider.future), 42);
       expect(container.read(provider), const AsyncValue.data(42));
       expect(root.getAllProviderElements(), isEmpty);
       expect(
@@ -286,7 +202,7 @@ void main() {
           isA<ProviderElementBase>()
               .having((e) => e.origin, 'origin', provider),
           isA<ProviderElementBase>()
-              .having((e) => e.origin, 'origin', provider.last),
+              .having((e) => e.origin, 'origin', provider.future),
           isA<ProviderElementBase>()
               .having((e) => e.origin, 'origin', provider.stream),
         ]),
@@ -312,7 +228,7 @@ void main() {
         container.read(provider), AsyncValue<int>.error(42, stackTrace: stack));
   });
 
-  group('.last', () {
+  group('.future', () {
     test(
         'throws StateError if the provider is disposed before a value was emitted',
         () async {
@@ -320,12 +236,12 @@ void main() {
         provider.overrideWithValue(const AsyncLoading()),
       ]);
 
-      final last = container.read(provider.last);
+      final future = container.read(provider.future);
 
       container.dispose();
 
       await expectLater(
-        last,
+        future,
         throwsA(
           isA<StateError>().having(
             (e) => e.message,
@@ -343,7 +259,7 @@ void main() {
         provider.overrideWithValue(const AsyncLoading()),
       ]);
 
-      var last = container.read(provider.last);
+      var future = container.read(provider.future);
 
       final error = Error();
 
@@ -351,10 +267,10 @@ void main() {
         provider.overrideWithValue(AsyncValue.error(error)),
       ]);
 
-      expect(container.read(provider.last), last);
+      expect(container.read(provider.future), future);
 
       // TODO(rrousselGit) test that the stacktrace is preserved
-      await expectLater(last, throwsA(error));
+      await expectLater(future, throwsA(error));
 
       final error2 = Error();
 
@@ -362,16 +278,16 @@ void main() {
         provider.overrideWithValue(const AsyncValue.loading()),
       ]);
 
-      last = container.read(provider.last);
+      future = container.read(provider.future);
 
       container.updateOverrides([
         provider.overrideWithValue(AsyncValue.error(error2)),
       ]);
 
-      expect(container.read(provider.last), last);
+      expect(container.read(provider.future), future);
 
       // TODO(rrousselGit) test that the stacktrace is preserved
-      await expectLater(last, throwsA(error2));
+      await expectLater(future, throwsA(error2));
     });
 
     test('supports loading then error then another error', () async {
@@ -379,7 +295,7 @@ void main() {
         provider.overrideWithValue(const AsyncLoading()),
       ]);
 
-      var last = container.read(provider.last);
+      var future = container.read(provider.future);
 
       final error = Error();
 
@@ -387,10 +303,10 @@ void main() {
         provider.overrideWithValue(AsyncValue.error(error)),
       ]);
 
-      expect(container.read(provider.last), last);
+      expect(container.read(provider.future), future);
 
       // TODO(rrousselGit) test that the stacktrace is preserved
-      await expectLater(last, throwsA(error));
+      await expectLater(future, throwsA(error));
 
       final error2 = Error();
 
@@ -399,10 +315,10 @@ void main() {
         provider.overrideWithValue(AsyncValue.error(error2)),
       ]);
 
-      last = container.read(provider.last);
+      future = container.read(provider.future);
 
       // TODO(rrousselGit) test that the stacktrace is preserved
-      await expectLater(last, throwsA(error2));
+      await expectLater(future, throwsA(error2));
     });
 
     test('supports loading then data then loading', () async {
@@ -410,27 +326,27 @@ void main() {
         provider.overrideWithValue(const AsyncLoading()),
       ]);
 
-      var last = container.read(provider.last);
+      var future = container.read(provider.future);
 
       container.updateOverrides([
         provider.overrideWithValue(const AsyncValue.data(42)),
       ]);
 
-      expect(container.read(provider.last), last);
-      await expectLater(last, completion(42));
+      expect(container.read(provider.future), future);
+      await expectLater(future, completion(42));
 
       container.updateOverrides([
         provider.overrideWithValue(const AsyncValue.loading()),
       ]);
 
-      last = container.read(provider.last);
+      future = container.read(provider.future);
 
       container.updateOverrides([
         provider.overrideWithValue(const AsyncValue.data(21)),
       ]);
 
-      expect(container.read(provider.last), last);
-      await expectLater(last, completion(21));
+      expect(container.read(provider.future), future);
+      await expectLater(future, completion(21));
     });
 
     test('supports loading then data then another data', () async {
@@ -438,23 +354,23 @@ void main() {
         provider.overrideWithValue(const AsyncLoading()),
       ]);
 
-      var last = container.read(provider.last);
+      var future = container.read(provider.future);
 
       container.updateOverrides([
         provider.overrideWithValue(const AsyncValue.data(42)),
       ]);
 
-      expect(container.read(provider.last), last);
-      await expectLater(last, completion(42));
+      expect(container.read(provider.future), future);
+      await expectLater(future, completion(42));
 
       // data without passing by an intermediary loading state
       container.updateOverrides([
         provider.overrideWithValue(const AsyncValue.data(21)),
       ]);
 
-      last = container.read(provider.last);
+      future = container.read(provider.future);
 
-      await expectLater(last, completion(21));
+      await expectLater(future, completion(21));
     });
   });
 
@@ -587,7 +503,8 @@ void main() {
     verifyNoMoreInteractions(listener);
   });
 
-  test('.last does not update dependents if the created future did not change',
+  test(
+      '.future does not update dependents if the created future did not change',
       () async {
     final dep = StateProvider((ref) => 0);
 
@@ -597,7 +514,7 @@ void main() {
     });
     final listener = Listener<Future<int>>();
 
-    container.listen(provider.last, listener, fireImmediately: true);
+    container.listen(provider.future, listener, fireImmediately: true);
 
     verifyOnly(listener, listener(any, any));
 
@@ -609,7 +526,7 @@ void main() {
     // No value were emitted, so the future will fail. Catching the error to
     // avoid false positive.
     // ignore: unawaited_futures, avoid_types_on_closure_parameters
-    container.read(provider.last).catchError((Object _) => 0);
+    container.read(provider.future).catchError((Object _) => 0);
   });
 
   group('overrideWithValue(T)', () {
@@ -878,10 +795,10 @@ void main() {
       );
     });
 
-    test('.name is the listened name.last', () {
+    test('.name is the listened name.future', () {
       expect(
-        StreamProvider<int>((ref) async* {}, name: 'hey').last.name,
-        'hey.last',
+        StreamProvider<int>((ref) async* {}, name: 'hey').future.name,
+        'hey.future',
       );
       expect(
         StreamProvider<int>((ref) async* {}).stream.name,
@@ -904,10 +821,12 @@ void main() {
       );
     });
 
-    test('.name is the listened name.last', () {
+    test('.name is the listened name.future', () {
       expect(
-        StreamProvider.autoDispose<int>((ref) async* {}, name: 'hey').last.name,
-        'hey.last',
+        StreamProvider.autoDispose<int>((ref) async* {}, name: 'hey')
+            .future
+            .name,
+        'hey.future',
       );
       expect(
         StreamProvider.autoDispose<int>((ref) async* {}).stream.name,
@@ -991,14 +910,14 @@ void main() {
     });
   });
 
-  group('StreamProvider.last', () {
+  group('StreamProvider.future', () {
     group('from StreamProvider', () {
       test('read currentValue before first value', () async {
         final container = createContainer();
         final controller = StreamController<int>();
         final provider = StreamProvider<int>((_) => controller.stream);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         controller.add(42);
 
@@ -1014,7 +933,7 @@ void main() {
 
         controller.add(42);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         await expectLater(future, completion(42));
 
@@ -1026,7 +945,7 @@ void main() {
         final controller = StreamController<int>();
         final provider = StreamProvider<int>((_) => controller.stream);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         controller.addError(42);
 
@@ -1042,7 +961,7 @@ void main() {
 
         controller.addError(42);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         await expectLater(future, throwsA(42));
 
@@ -1057,7 +976,7 @@ void main() {
           provider.overrideWithValue(const AsyncValue.loading()),
         ]);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         container.updateOverrides([
           provider.overrideWithValue(const AsyncValue.data(42)),
@@ -1076,7 +995,7 @@ void main() {
           provider.overrideWithValue(const AsyncValue.data(42)),
         ]);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         await expectLater(future, completion(42));
       });
@@ -1087,7 +1006,7 @@ void main() {
           provider.overrideWithValue(const AsyncValue.loading()),
         ]);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         container.updateOverrides([
           provider.overrideWithValue(const AsyncValue.error(42)),
@@ -1106,7 +1025,7 @@ void main() {
           provider.overrideWithValue(const AsyncValue.error(42)),
         ]);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         await expectLater(future, throwsA(42));
       });
@@ -1117,7 +1036,7 @@ void main() {
           provider.overrideWithValue(const AsyncValue.data(42)),
         ]);
 
-        final future = container.read(provider.last);
+        final future = container.read(provider.future);
 
         await expectLater(future, completion(42));
       });